+++ plugins/ocp/ocp-utils.h
@@ -17,10 +17,10 @@ extern const unsigned char ocp_uuid[NVME_UUID_LEN];
  * @dev:	nvme device
  * @index:	integer pointer to here to save the index
  *
- * Return: Zero if nvme device has UUID list log page, or positive result of get uuid list
+ * Return: Zero if nvme device has UUID list identify page, or positive result of get uuid list
  *         or negative POSIX error code otherwise.
  */
-int ocp_get_uuid_index(struct nvme_dev *dev, int *index);
+int ocp_get_uuid_index(struct nvme_dev *dev, __u8 *index);
 
 /**
  * ocp_find_uuid_index() - Find OCP UUID index in UUID list
@@ -29,4 +29,4 @@ int ocp_get_uuid_index(struct nvme_dev *dev, int *index);
  *
  * Return: Zero if nvme device has UUID list log page, Negative POSIX error code otherwise.
  */
-int ocp_find_uuid_index(struct nvme_id_uuid_list *uuid_list, int *index);
+int ocp_find_uuid_index(struct nvme_id_uuid_list *uuid_list, __u8 *index);
